♻️ Refactor Code / not use DOM API
